+++ xolotl/xolotlCore/displacement/IDisplacementHandler.h
@@ -8,8 +8,8 @@
 namespace xolotlCore {
 
 /**
- * Realizations of this interface are responsible for handling the incident (incoming)
- * flux calculations.
+ * Realizations of this interface are responsible for handling the initial
+ * displacement calculations.
  */
 class IDisplacementHandler {
 
@@ -18,7 +18,7 @@ public:
 	virtual ~IDisplacementHandler() { }
 
 	/**
-	 * Compute and store the incident flux values at each grid point.
+	 * Compute and store the initial vacancy values at each grid point.
 	 *
 	 * @param network The reaction network
 	 * @param nx The total number of grid points that will be used on the x axis
@@ -28,7 +28,7 @@ public:
 			int nx, double hx) = 0;
 
 	/**
-	 * This operation returns the incident flux vector.
+	 * This operation returns the initial displacement vector.
 	 *
 	 * @param currentTime The time
 	 * @return The incident flux vector
@@ -36,42 +36,40 @@ public:
 	virtual std::vector<double> getInitialDisplacementVec() = 0;
 
 	/**
-	 * This operation returns the index of the cluster that is irradiating
-	 * the material.
+	 * This operation returns the index of the vacancy cluster.
 	 *
 	 * @return The index of the incident flux cluster
 	 */
 	virtual int getInitialDisplacementClusterIndex() = 0;
 
 	/**
-	 * This operation sets the factor to change the intensity of the flux.
+	 * This operation sets the factor to change the intensity of the krypton fluence.
 	 *
 	 * @param flux The flux intensity
 	 */
 	virtual void setKrFluenceAmplitude(double krFluence) = 0;
 
 	/**
-	 * This operation gets the factor that changes the flux
-	 * intensity/amplitude.
+	 * This operation gets the factor to change the intensity of the krypton fluence.
 	 *
 	 * @return The flux amplitude
 	 */
 	virtual double getKrFluenceAmplitude() const = 0;
 
 	/**
-	 * This operation sets the factor to change the intensity of the flux.
+	 * This operation sets the factor to change the threshold energy.
 	 *
 	 * @param flux The flux intensity
 	 */
-	virtual void setDispEnergy(double thresholdEnergy) = 0;
+	virtual void setDispEnergy(int thresholdEnergy) = 0;
 
 	/**
-	 * This operation gets the factor that changes the flux
+	 * This operation gets the factor that changes the threshold energy.
 	 * intensity/amplitude.
 	 *
 	 * @return The flux amplitude
 	 */
-	virtual double getDispEnergy() const = 0;
+	virtual int getDispEnergy() const = 0;
 
 }; //end class IFluxHandler
